requirements.txt
python-telegram-bot==21.6
replicate>=1.0.0,<2.0.0
requests==2.32.3
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
python-dotenv>=1.0.1